[WAYANG-#8] Change structure of Graph
