+++ python/src/pywy/old_graph/visitant.py
